+++ src/components/pages/common/upload-links-popup/types.tsx
@@ -3,7 +3,7 @@ import { TLinkDecrypted } from 'types'
 export type TProps = {
   onClose: () => void,
   onSubmit: (links: TLinkDecrypted[]) => void,
-  quantity: number,
+  quantity?: number,
   loader: number,
   loading: boolean
 }