+++ src/components/pages/index.js
@@ -12,9 +12,13 @@ import QRDownload from './qr-download/index'
 import CampaignsCreateNew from './campaigns-create-new'
 import Main from './main'
 import QRCreate from './qr-create'
+import Dispensers from './dispensers'
+import DispenserCreate from './dispenser-create'
+import Dispenser from './dispenser'
 
 export {
-  Page, 
+  Page,
+  Dispenser,
   NotFound,
   Campaigns,
   CampaignsDetails,
@@ -27,5 +31,7 @@ export {
   QR,
   QRDownload,
   Main,
-  QRCreate
+  QRCreate,
+  Dispensers,
+  DispenserCreate
 }