+++ src/components/pages/qr/components/index.tsx
@@ -1,9 +1,5 @@
 import QuantityPopup from "./quantity-popup"
-import LinksPopup from "./links-popup"
-import DownloadPopup from "./download-popup"
 
 export {
-  QuantityPopup,
-  LinksPopup,
-  DownloadPopup
+  QuantityPopup
 }